checker: move Guard to its own file
Guard is an independent axis, not an aspect of Effect; one concept per
file, matching level.go and change.go.
